+++ R/RFBCoptions.R
@@ -21,12 +21,10 @@
 #' \item{\code{scaleBarLength}: }{A single numeric that represents the actual length of the scale-bar. Ignored if \code{scaleBar=FALSE}. Defaults to \code{NULL}; thus, the user must enter a value if \code{scaleBar=TRUE}. Used in \code{\link{digitizeRadii}}.}
 #' \item{\code{col.scaleBar}: }{The color of the scale-bar line if \code{scalebar=TRUE}. Defaults to \code{"red"}. Used in \code{\link{digitizeRadii}}, \code{\link{showDigitizedImage}}, and \code{\link{findScalingFactor}}.}
 #' \item{\code{lwd.scaleBar}: }{The line width of the scale-bar line if \code{scalebar=TRUE}. Defaults to \code{2}. Used in \code{\link{digitizeRadii}}, \code{\link{showDigitizedImage}}, and \code{\link{findScalingFactor}}.}
-#' \item{\code{addTransect}: }{A single logical that indicates whether the points selected at the structure center and margin should be connected to form a linear transect on which annuli will be marked. Defaults to \code{TRUE}. Used in \code{\link{digitizeRadii}}.}
-#' \item{\code{snap2Transect}: }{A single logical that indicates whether the coordinates of the selected points that represent annuli should be moved to fall exactly on the transect from the structure center to margin. If \code{TRUE} then the points will be moved perpendicularly to the transect. If \code{FALSE} then the points will be where the user selected them. Defaults to \code{TRUE}. Used in \code{\link{digitizeRadii}} but the result will not be seen until visualizing the selected points with \code{\link{showDigitizedImage}}.}
-#' 
-#' \item{\code{showTransect}: }{A single logical that indicates whether the points selected at the structure center and margin should be connected to form a linear transect. Defaults to \code{TRUE}. Used in \code{\link{showDigitizedImage}}.}
-#' \item{\code{col.transect}: }{The color of the transect line if \code{addTransect=TRUE} in \code{\link{digitizeRadii}} or \code{showTransect=TRUE} in \code{\link{showDigitizedImage}}. Defaults to \code{"yellow"}.}
-#' \item{\code{lwd.transect}: }{The width of the transect line if \code{addTransect=TRUE} in \code{\link{digitizeRadii}} or \code{showTransect=TRUE} in \code{\link{showDigitizedImage}}. Defaults to \code{1}.}
+#' \item{\code{showTransect}: }{A single logical that indicates whether the transection between the points selected at the structure center and margin should be shown. Defaults to \code{TRUE}. Used in \code{\link{digitizeRadii}} and  \code{\link{showDigitizedImage}}.}
+#' \item{\code{snap2Transect}: }{A single logical that indicates whether the coordinates of the selected points that represent annuli should be moved to fall exactly on the transect from the structure center to margin. If \code{TRUE} then the points will be moved perpendicularly to the transect (and the original user-selected point will not be seen on the image). If \code{FALSE} then the points will be where the user selected them. Defaults to \code{TRUE}. Used in \code{\link{digitizeRadii}}.}
+#' \item{\code{col.transect}: }{The color of the transect line if \code{showTransect=TRUE} in \code{\link{digitizeRadii}} or \code{showTransect=TRUE} in \code{\link{showDigitizedImage}}. Defaults to \code{"yellow"}.}
+#' \item{\code{lwd.transect}: }{The width of the transect line if \code{showTransect=TRUE} in \code{\link{digitizeRadii}} or \code{showTransect=TRUE} in \code{\link{showDigitizedImage}}. Defaults to \code{1}.}
 #' \item{\code{pch.sel}: }{The plotting character of points for selected annuli in \code{\link{digitizeRadii}}. Defaults to \code{3} (a \dQuote{cross hairs}).}
 #' \item{\code{col.sel}: }{The color of points for selected annuli in \code{\link{digitizeRadii}}. Defaults to \code{"red"}.}
 #' \item{\code{cex.sel}: }{The character expansion value of points for selected annuli in \code{\link{digitizeRadii}}. Defaults to \code{1}.}
@@ -56,11 +54,11 @@
 #' RFBCoptions()
 #' 
 #' ## Show how to see and set one option
-#' RFBCoptions()$addTransect
-#' RFBCoptions(addTransect=FALSE)
-#' RFBCoptions()$addTransect
+#' RFBCoptions()$showTransect
+#' RFBCoptions(showTransect=FALSE)
+#' RFBCoptions()$showTransect
 #' RFBCoptions(reset=TRUE)
-#' RFBCoptions()$addTransect
+#' RFBCoptions()$showTransect
 #' 
 #' ## Multiple options can also be set at once
 #' RFBCoptions(pch.show=3,col.show="blue",cex.show=3)
@@ -79,7 +77,7 @@ iRFBCopts <- settings::options_manager(reading=NULL,description=NULL,
                 sepWindow=TRUE,windowSize=7,popID=FALSE,
                 scalingFactor=1,scaleBar=FALSE,scaleBarLength=NULL,
                 col.scaleBar="red",lwd.scaleBar=2,
-                addTransect=TRUE,snap2Transect=TRUE,showTransect=TRUE,
+                showTransect=TRUE,snap2Transect=TRUE,
                 col.transect="yellow",lwd.transect=1,
                 pch.sel=3,col.sel="red",cex.sel=1,
                 pch.show=19,col.show="red",cex.show=1,
@@ -90,9 +88,8 @@ iRFBCopts <- settings::options_manager(reading=NULL,description=NULL,
                 scalingFactor=settings::inrange(min=1e-10,max=Inf),
                 scaleBar=settings::inlist(TRUE,FALSE),
                 lwd.scaleBar=settings::inrange(min=1,max=10),
-                addTransect=settings::inlist(TRUE,FALSE),
-                snap2Transect=settings::inlist(TRUE,FALSE),
                 showTransect=settings::inlist(TRUE,FALSE),
+                snap2Transect=settings::inlist(TRUE,FALSE),
                 lwd.transect=settings::inrange(min=1,max=10),
                 cex.sel=settings::inrange(min=0.1,max=10),
                 cex.show=settings::inrange(min=0.1,max=10),