+++ R/digitizeRadii.R
@@ -16,7 +16,7 @@
 #' @param col.scaleBar See details in \code{\link{RFBCoptions}}.
 #' @param lwd.scaleBar See details in \code{\link{RFBCoptions}}.
 #' @param scalingFactor See details in \code{\link{RFBCoptions}}.
-#' @param addTransect See details in \code{\link{RFBCoptions}}.
+#' @param showTransect See details in \code{\link{RFBCoptions}}.
 #' @param snap2Transect See details in \code{\link{RFBCoptions}}.
 #' @param col.transect See details in \code{\link{RFBCoptions}}.
 #' @param lwd.transect See details in \code{\link{RFBCoptions}}.
@@ -41,8 +41,10 @@
 #'   \item{\code{basenm}: }{The filename given in \code{img} without the path.}
 #'   \item{\code{dirnm}: }{The directory in which \code{img} was located.}
 #'   \item{\code{datanm}: }{The filename that contains the RData object.}
-#'   \item{\code{pts}: }{A data.frame that contains the \code{x} and \code{y} coordinates on the image for the selected annuli.}
+#'   \item{\code{pts}: }{A data.frame that contains the \code{x} and \code{y} coordinates on the image for the selected annuli. These points may be \dQuote{snapped} to the transect if \code{snap2Transect==TRUE}.}
+#'   \item{\code{orig.pts}: }{A data.frame that contains the \dQuote{original} \code{x} and \code{y} coordinates on the image for the selected annuli. If \code{snap2Transect=FALSE} then these are the same as the points in \code{pts}. If \code{snap2Transect=TRUE} then these are the originally selected points and will be different then the points in \code{pts}.}
 #'   \item{\code{radii}: }{A data.frome that contains the unique \code{id}, the \code{reading} code, the age-at-capture in \code{agecap}, the annulus number in \code{ann}, the radial measurements in \code{rad}, and the radial measurement at capture in \code{radcap}.}
+#'   \item{\code{snap2Transect}: }{The logical from \code{snap2Transect} that identified whether the selected points were \dQuote{"snapped"} to the transect or not.}
 #'   \item{\code{sfSource}: }{A character string that identifies whether the scaling factor was \code{"Provided"} through the \code{scalingFactor} argument or derived from a \code{"scaleBar"}.}
 #'   \item{\code{sbPts}: }{A data.frame of \code{x} and \code{y} coordinates for the endpoints of the scale-bar if the scaling factor was derived from a scale-bar.}
 #'   \item{\code{sbLength}: }{A single numeric that is the known (actual) length of the scale-bar if the scaling factor was derived from a scale-bar.}
@@ -63,7 +65,7 @@ digitizeRadii <- function(img,id,reading,suffix,
                           description,edgeIsAnnulus,popID,
                           sepWindow,windowSize,scaleBar,
                           scaleBarLength,col.scaleBar,lwd.scaleBar,
-                          scalingFactor,addTransect,snap2Transect,
+                          scalingFactor,showTransect,snap2Transect,
                           col.transect,lwd.transect,
                           pch.sel,col.sel,cex.sel,
                           showInfo,pos.info,cex.info,col.info) {
@@ -82,7 +84,7 @@ digitizeRadii <- function(img,id,reading,suffix,
   if (missing(col.scaleBar)) col.scaleBar <- iGetopt("col.scaleBar")
   if (missing(lwd.scaleBar)) lwd.scaleBar <- iGetopt("lwd.scaleBar")
   if (missing(scalingFactor)) scalingFactor <- iGetopt("scalingFactor")
-  if (missing(addTransect)) addTransect<- iGetopt("addTransect")
+  if (missing(showTransect)) showTransect<- iGetopt("showTransect")
   if (missing(snap2Transect)) snap2Transect<- iGetopt("snap2Transect")
   if (missing(col.transect)) col.transect <- iGetopt("col.transect")
   if (missing(lwd.transect)) lwd.transect <- iGetopt("lwd.transect")
@@ -107,17 +109,15 @@ digitizeRadii <- function(img,id,reading,suffix,
                            col.scaleBar,lwd.scaleBar,
                            pixW2H=windowInfo$pixW2H)
   ## User selects annuli on the image
-  pts <- iSelectAnnuli(pch.pts=pch.sel,col.pts=col.sel,
-                       cex.pts=cex.sel,addTransect=addTransect,
-                       col.trans=col.transect,lwd.trans=lwd.transect)
-  ## If asked, "snap" the selected points to the transect
-  pts <- iSnap2Transect(pts)
+  pts <- iSelectAnnuli(pch.pts=pch.sel,col.pts=col.sel,cex.pts=cex.sel,
+                       showTransect=showTransect,snap2Transect=snap2Transect,
+                       col.transect=col.transect,lwd.transect=lwd.transect)
   ## Converts the selected points to radial measurements
   dat <- iProcessAnnuli(inames,pts,id,reading,suffix,description,
                         edgeIsAnnulus,SF$scalingFactor,
                         pixW2H=windowInfo$pixW2H)
   ## Add windowSize and scaling factor information to dat list
-  dat <- c(dat,SF,windowInfo)
+  dat <- c(dat,list(snap2Transect=snap2Transect),SF,windowInfo)
   ## Write the dat object to R object filename in the working directory.
   save(dat,file=paste0(inames$dirnm,"/",dat$datanm))
   message("** All results written to ",dat$datanm)